moved Login and Register components to components/logon
