+++ src/routes/Routes.js
@@ -1,11 +1,11 @@
 import React from "react";
 import { Route } from "react-router";
 import PrivateRoute from "./PrivateRoute";
-import Login from "../components/Login";
-import Register from "../components/Register";
+import Login from "../components/logon/Login";
+import Register from "../components/logon/Register";
 import HelloWorld from "../components/HelloWorld";
 import StrainForm from "../components/StrainForm";
-import { Ailments } from "../components/Intake/Ailments";
+import { Ailments } from "../components/intake/Ailments";
 
 const Routes = () => {
   return (
